chore: apply deferred review cleanups from PR #15
- ArrowRightCircle: one shared component (was byte-identical in both
  boards, down to the comment)
- usePrefersReducedMotion: shared @utils hook with a legacy addListener
  fallback — MediaQueryList.addEventListener throws on Safari ≤13 and an
  uncaught effect error unmounts the whole island
- --ease-spring token on :root replaces 14 copies of the house overshoot
  curve across 6 files (flagged in three reviews running); TSX inline
  styles reference it via var() too
- Band custom props declared once per path: the viewport-approximation
  fallbacks now live inside @supports not (container-type), so the two
  @container >= re-assert blocks are gone and the paths can't drift
  (equivalence verified in review)
- Pull cord hidden until hydration adds pull-container--ready: the
  handlers are React-bound, so the SSR'd cord was a visible dead control
  through the idle window (and forever with JS off); older posts stay
  reachable JS-off via the paginated tag pages

Verified in the build: one bezier literal left (the token), @supports-not
present with >= blocks gone, cord SSR'd hidden and gaining --ready after
hydration (DOM dump), shared icon in both island bundles, renders
unchanged at both widths.
