+++ src/components/RisoNav.tsx
@@ -66,7 +66,7 @@ export default function RisoNav({
             className="after:absolute after:-inset-4 after:content-['']"
           />
           <div
-            className="dust-note pointer-events-none absolute left-[calc(100%+12px)] top-1/2 z-[100] hidden min-w-[200px] max-w-[300px] -translate-x-[10px] -translate-y-1/2 rotate-[3deg] border border-dashed border-black/20 bg-[var(--yellow)] px-[10px] py-1.5 text-center font-mono text-sm font-bold uppercase leading-tight tracking-wide text-[var(--black)] opacity-0 shadow-[2px_3px_6px_rgba(0,0,0,0.2)] transition-[opacity_0.2s_ease,transform_0.3s_cubic-bezier(0.34,1.56,0.64,1)] [clip-path:polygon(0%_0%,100%_0%,98%_100%,2%_95%)] group-hover:-translate-y-1/2 group-hover:translate-x-[5px] group-hover:-rotate-2 group-hover:opacity-100 motion-reduce:transition-[opacity_0.2s_ease,transform_0.2s_ease] transform-gpu backface-hidden md:block"
+            className="dust-note pointer-events-none absolute left-[calc(100%+12px)] top-1/2 z-[100] hidden min-w-[200px] max-w-[300px] -translate-x-[10px] -translate-y-1/2 rotate-[3deg] border border-dashed border-black/20 bg-[var(--yellow)] px-[10px] py-1.5 text-center font-mono text-sm font-bold uppercase leading-tight tracking-wide text-[var(--black)] opacity-0 shadow-[2px_3px_6px_rgba(0,0,0,0.2)] transition-[opacity_0.2s_ease,transform_0.3s_var(--ease-spring)] [clip-path:polygon(0%_0%,100%_0%,98%_100%,2%_95%)] group-hover:-translate-y-1/2 group-hover:translate-x-[5px] group-hover:-rotate-2 group-hover:opacity-100 motion-reduce:transition-[opacity_0.2s_ease,transform_0.2s_ease] transform-gpu backface-hidden md:block"
             aria-hidden="true"
           >
             Pardon the dust, <br />
